multi: Update LDK to 0.0.113 / 0.21 + fix
Like 0.0.112, 0.0.113 is broken; LDK panics with 50% probability if we
use random nanos:

https://github.com/lightningdevkit/rust-lightning/pull/1935

This PR updates LDK past 0.0.113 to the commit that merges the fix.

+++ common/src/ln/channel.rs
@@ -71,7 +71,7 @@ pub struct LxChannelDetails {
     pub inbound_scid_alias: Option<u64>,
     pub channel_value_satoshis: u64,
     pub unspendable_punishment_reserve: Option<u64>,
-    pub user_channel_id: u64,
+    pub user_channel_id: u128,
     pub balance_msat: u64,
     pub outbound_capacity_msat: u64,
     pub next_outbound_htlc_limit_msat: u64,
